Refactor component decorators into module
+ cleanup controllers for app and default layout.

+ created stubs for pending Prop, Watch and Emit decorators.
