--- src/layouts/default/default.ts
@@ -1,5 +1,0 @@
-export default {
-  setup() {
-    return {};
-  },
-};
