--- src/layouts/default/index.ts
@@ -1,1 +1,0 @@
-export { default } from './default.vue';
