+++ src/modules/component/index.ts
@@ -1,0 +1,5 @@
+export * from './base';
+export * from './hooks';
+export * from './prop';
+export * from './emit';
+export * from './watch';
