+++ src/modules/core/index.ts
@@ -1,3 +1,2 @@
 export * from './font-awesome-plugin';
 export * from './i18n-plugin';
-export * from './component';
